Added code to translate morse and convert it into blinking of the LED
I made this code in another REPO to test, but it now works. There is the controller to communicate with the Arduino (from Pythondaq) and some code to translate words to morse (which put in controllers for the time being). I also made a simple controller that turns the dots and stripes into short and long pulses of the LED.